Split new functionalities into a new file cart.py
